requirements.txt
beautifulsoup4==4.8.0
Flask==1.1.1
flask-paginate==0.5.3
flask-talisman==0.7.0
gunicorn==19.9.0
Jinja2==2.10.1
mongoengine==0.18.2
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3